Add files via upload
This is a program that allows you to retrieve these tracks from GPS tracks made with an iWatch connected watch; these traces can thus be used in very practical mapping software for planning certain mountain routes in particular.
Because Apple maintains a certain tendency to restrict the functioning of its products. Thus, if it is possible to extract a GPS trace from its iWatch via the unavoidable means of its iPhone, this trace will not be usable with non-Apple software. Because the format of the trace recovered in the iWatch is typically proprietary and is presented in a format specific to Apple.
The software proposed here therefore works to convert this Apple track into an official GPX format widely used by current mapping systems such as the Garmin Basecamp and others.